% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_templates.R
\name{available_templates}
\alias{available_templates}
\title{Get Named Vector of Available Post Templates}
\usage{
available_templates()
}
\description{
Call this function to get a named vector of available templates for new
posts. Post templates must follow the following path convention:
\code{inst/rmarkdown/templates/template_name/skeleton/skeleton.Rmd}.
}
