% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{create_post_from_template_addin}
\alias{create_post_from_template_addin}
\title{Create Post from Template}
\usage{
create_post_from_template_addin()
}
\description{
Call this function as an addin to interactively create a new post from an
R Markdown template.
}
