% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_talk.R
\name{create_talk}
\alias{create_talk}
\title{Create a talk page with icon links}
\usage{
create_talk(
  title,
  collection = "talks",
  author = "auto",
  slug = "auto",
  date = Sys.Date(),
  date_prefix = date,
  draft = FALSE,
  open = TRUE
)
}
\arguments{
\item{title}{Post title}

\item{collection}{collection to create the post within (defaults to "talks")}

\item{author}{Post author. Automatically drawn from previous post if not provided.}

\item{slug}{Post slug (directory name). Automatically computed from title if not provided}

\item{date}{Post date (defaults to current date)}

\item{date_prefix}{Date prefix for post slug (preserves chronological order for talks within the filesystem). Pass \code{NULL} for no date prefix}

\item{draft}{Mark the post as a \code{draft} (don't include in the article listing)}

\item{open}{logical. Open the created file?}
}
\description{
This function wraps \code{\link[distill:create_post]{distill::create_post()}}, but defaults to the \verb{_talk} collection and pre-populates the post with \code{\link[=icon_link]{icon_link()}} functions to provide buttons for slides, materials and project. These appear in a code chunk and can be edited. The \code{edit} option in the call to \code{\link[distill:create_post]{distill::create_post()}} must be set to \code{FALSE}, so an \code{open} argument is provided instead that serves the same purpose.
}
\note{
This function must be called from with a working directory that is within
a Distill website.
}
\examples{
\dontrun{
library(distilltools)
create_talk("My Talk")
}
}
\author{
Eric Ekholm and Ella Kaye
}
