% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{make_icon}
\alias{make_icon}
\alias{icon_link}
\title{Link "buttons" with icons and text}
\usage{
make_icon(icon)

icon_link(icon = NULL, text = NULL, url = NULL)
}
\arguments{
\item{icon}{The name of the icon. For Font Awesome icons, the name should correspond to current Version 5 long names (e.g. \code{"fab fa-github"}). \href{https://jpswalsh.github.io/academicons/}{Academicons} can also be used and styled the same way (e.g. \code{"ai ai-google-scholar"}) but require a site header. See \href{https://www.jhelvy.com/posts/2021-03-25-customizing-distill-with-htmltools-and-css/#side-note-on-academic-icons}{here} for details.}

\item{text}{A string of the text to appear on the button}

\item{url}{A url for where the button should link to}
}
\value{
For \code{make_icon}, a \code{shiny.tag} with the HTML \verb{<i class = "icon"></i>}. For \code{icon_link}, a \code{shiny.tag} with the HTML \verb{<a href=url class="icon-link"><i class=icon></i> text</a>}
}
\description{
Easily insert "buttons" with an icon and text into .Rmd documents
}
\details{
Note that it will be necessary to add some CSS to in order to make the "buttons" look like buttons. For the \code{icon_link} function, this can be done by styling the \code{icon-link} class. At a minimum, a border property should be set. For a \code{distill} website, we recommend using the \code{distill::create_theme} function and writing additional CSS at the bottom of that file. See \href{https://rstudio.github.io/distill/website.html#theming}{here} for details of how to implement that. See \href{https://github.com/jhelvy/jhelvy.com/blob/master/css/jhelvy.css}{here} and \href{https://github.com/EllaKaye/ellakaye-distill/blob/main/emk_theme.css}{here} for examples.
}
\examples{
make_icon("r-project")
icon_link("fab fa-github", "materials", "https://github.com/USER/REPO")
}
\author{
John Paul Helveston and Ella Kaye
}
