% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_templates.R
\name{available_templates}
\alias{available_templates}
\title{Get Named Vector of Available Post Templates}
\usage{
available_templates()
}
\description{
Get a named vector of available R Markdown templates for new posts.
}
\details{
By default \code{distilltools} looks for R Markdown post templates in
\verb{./inst/rmarkdown/templates/}. In order to be recognized, templates within
this path must be structured as: \code{template_name/skeleton/skeleton.Rmd}.

The default path can be changed with
\code{options("distilltools.templates.path" = "custom/template/path/")}.
For the change to persist between R sessions this code should be placed
in a project local \code{.Rprofile}. Use \code{usethis::edit_r_profile("project")} to
create or edit this file.
}
\seealso{
\code{distilltools:create_post_from_template()}
}
