% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_templates.R
\name{available_templates}
\alias{available_templates}
\title{Get Named Vector of Available Post Templates}
\usage{
available_templates()
}
\description{
Get a named vector of available R Markdown templates for new posts.
}
\details{
By default \code{available_templates()} looks for R Markdown post templates in
\code{./templates}. The template file name will be used to name the template. For
example, \code{./templates/my_template.Rmd} will be named "my_template". This
name can be used to access the path for "my_template" (see examples), and
will also be displayed in the list of available templates in the "Create
post from template" RStudio addin.

The default directory \code{available_templates()} looks for R Markdown post
templates in can be changed with
\code{options("distilltools.templates.path" = "custom/templates/path/")}.
For the change to persist between R sessions this code should be placed
in a project local \code{.Rprofile}. Use \code{usethis::edit_r_profile("project")} to
create or edit this file.

In addition to local templates, \code{available_templates()} will always make
available a default template named "Default" that matches the one from
\code{distill::create_post()}. This template can be selected in the "Create
post from template" RStudio addin as well.
}
\examples{
# Return the path of a template using its name
available_templates()[["Default"]]

}
\seealso{
\verb{[distill::create_post()]}, \verb{[distilltools::create_post_from_template()]}
}
