% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_post_from_template.R
\name{create_post_from_template}
\alias{create_post_from_template}
\title{Create a new blog post from a template}
\usage{
create_post_from_template(
  path,
  title,
  collection = "posts",
  author = "auto",
  slug = "auto",
  date = Sys.Date(),
  date_prefix = date,
  draft = FALSE,
  edit = interactive()
)
}
\arguments{
\item{path}{File path to .Rmd template for post}

\item{title}{Post title. If there is a title in the template, this will override it.}

\item{collection}{Collection to create the post within (defaults to "posts")}

\item{author}{Post author. If not provided, it will be automatically drawn from template. If not in template, it will be automatically drawn from previous post.}

\item{slug}{Post slug (directory name). Automatically computed from title if not provided.}

\item{date}{Post date (defaults to current date)}

\item{date_prefix}{Date prefix for post slug (preserves chronological order for posts
within the filesystem). Pass \code{NULL} for no date prefix.}

\item{draft}{Mark the post as a \code{draft} (don't include it in the article listing).}

\item{edit}{Open the post in an editor after creating it.}
}
\description{
This function is designed to work as close as possible to \code{\link[distill:create_post]{distill::create_post()}} (and draws on a lot of code from that function). The main difference is the compolusary argument \code{path} which gives the path of the template to create the post from. All other arguments work the same as \code{\link[distill:create_post]{distill::create_post()}} (unless specified in Arguments).
}
\note{
This function must be called from with a working directory that is within a Distill website.

The output for a post must be \code{distill::distill_article}. If there is no output key in the provided template, this will be added to the yaml. If there is an \code{output} specified in the template yaml, it must be \code{distill::distill_article}, otherwise \code{create_post_from_template} will throw an error.

Unlike \code{\link[distill:create_post]{distill::create_post()}}, \code{create_post_from_template} doesn't automatically provide a \code{description} key in the yaml.

We recommend the use of \code{\link[here:here]{here::here()}} in specifying the path to the template
}
\examples{
\dontrun{
library(distilltools)
# .Rmd templates stored in "templates" directory
create_post(here::here("templates", "post-template.Rmd"))
}

}
